+++ src/Contract/RouteOverviewActionInterface.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Storage\Base\Contract;
 
+use Heptacom\HeptaConnect\Storage\Base\Exception\InvalidOverviewCriteriaException;
 use Heptacom\HeptaConnect\Storage\Base\Exception\UnsupportedStorageKeyException;
 
 interface RouteOverviewActionInterface
@@ -10,6 +11,7 @@ interface RouteOverviewActionInterface
     /**
      * @return iterable<RouteOverviewResult>
      *
+     * @throws InvalidOverviewCriteriaException
      * @throws UnsupportedStorageKeyException
      */
     public function overview(RouteOverviewCriteria $criteria): iterable;